pyproject.toml
[project]
name = "pdam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "d-geolib>=2.8.0",
    "pydantic>=2.12.5",
    "pyshp>=3.0.3",
    "shapely>=2.1.2",
]
